agents/short-bot/deployment/requirements_local.txt
absl-py==2.2.2
-e git+https://github.com/lalehaa/agent-demo.git@38ebeb4ccc04371d81fb155ae5723fab8f6f5d3c#egg=adk_short_bot&subdirectory=agents/adk_short_bot
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.5.2
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.0
cloudpickle==3.1.1
crashtest==0.4.1
cryptography==44.0.3
db-dtypes==1.4.3
Deprecated==1.2.18
distlib==0.3.9
docstring_parser==0.16
dulwich==0.22.8
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
google-adk==0.1.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.92.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.15.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.6.1
iniconfig==2.1.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
keyring==25.6.0
mcp==1.8.1
more-itertools==10.7.0
msgpack==1.1.0
numpy==2.2.5
opentelemetry-api==1.33.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
packaging==25.0
pandas==2.2.3
pbs-installer==2025.4.9
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.5.0
poetry==2.1.3
poetry-core==2.1.3
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
SecretStorage==3.3.3
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlglot==26.17.1
sse-starlette==2.3.5
starlette==0.46.2
tabulate==0.9.0
tomlkit==0.13.2
trove-classifiers==2025.5.9.12
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.2
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
